activity2concept; useractivity

+++ ums/trunk/lib/model/doctrine/base/BaseUserModel.class.php
@@ -10,12 +10,12 @@ abstract class BaseUserModel extends sfDoctrineRecord
         $this->setTableName('user_model');
         $this->hasColumn('concept_id', 'integer', null, array('type' => 'integer'));
         $this->hasColumn('user_id', 'integer', null, array('type' => 'integer'));
-        $this->hasColumn('bloom_evaluation', 'integer', null, array('type' => 'integer'));
-        $this->hasColumn('bloom_synthesis', 'integer', null, array('type' => 'integer'));
-        $this->hasColumn('bloom_analysis', 'integer', null, array('type' => 'integer'));
-        $this->hasColumn('bloom_application', 'integer', null, array('type' => 'integer'));
-        $this->hasColumn('bloom_understanding', 'integer', null, array('type' => 'integer'));
-        $this->hasColumn('bloom_knowledge', 'integer', null, array('type' => 'integer'));
+        $this->hasColumn('bloom_evaluation', 'float', null, array('type' => 'float'));
+        $this->hasColumn('bloom_synthesis', 'float', null, array('type' => 'float'));
+        $this->hasColumn('bloom_analysis', 'float', null, array('type' => 'float'));
+        $this->hasColumn('bloom_application', 'float', null, array('type' => 'float'));
+        $this->hasColumn('bloom_understanding', 'float', null, array('type' => 'float'));
+        $this->hasColumn('bloom_knowledge', 'float', null, array('type' => 'float'));
     }
 
     public function setUp()
